Moved style package to the jaraccount package
