+++ app/src/main/java/com/example/jarchess/jaraccount/styles/chessboard/MarbleChessboardStyle.java
@@ -1,4 +1,4 @@
-package com.example.jarchess.match.styles.chessboard;
+package com.example.jarchess.jaraccount.styles.chessboard;
 
 import com.example.jarchess.R;
 import com.example.jarchess.match.Coordinate;
