+++ app/src/main/java/com/example/jarchess/match/participant/AIOpponent.java
@@ -16,8 +16,8 @@ import com.example.jarchess.match.history.MatchHistory;
 import com.example.jarchess.match.move.Move;
 import com.example.jarchess.match.pieces.Piece;
 import com.example.jarchess.match.pieces.PromotionChoice;
-import com.example.jarchess.match.styles.avatar.AIAvatarStyle;
-import com.example.jarchess.match.styles.avatar.AvatarStyle;
+import com.example.jarchess.jaraccount.styles.avatar.AIAvatarStyle;
+import com.example.jarchess.jaraccount.styles.avatar.AvatarStyle;
 import com.example.jarchess.match.turn.Turn;
 import com.example.jarchess.testmode.TestableCurrentTime;
 import com.example.jarchess.testmode.TestableRandom;
