+++ app/src/main/java/com/example/jarchess/match/view/CapturedPiecesView.java
@@ -8,7 +8,7 @@ import com.example.jarchess.match.ChessColor;
 import com.example.jarchess.match.Coordinate;
 import com.example.jarchess.match.activity.MatchActivity;
 import com.example.jarchess.match.pieces.Piece;
-import com.example.jarchess.match.styles.chesspiece.ChesspieceStyle;
+import com.example.jarchess.jaraccount.styles.chesspiece.ChesspieceStyle;
 
 import static com.example.jarchess.match.ChessColor.BLACK;
 import static com.example.jarchess.match.ChessColor.WHITE;
