+++ app/src/main/java/com/example/jarchess/match/view/MatchView.java
@@ -25,8 +25,8 @@ import com.example.jarchess.match.move.PieceMovement;
 import com.example.jarchess.match.participant.MatchParticipant;
 import com.example.jarchess.match.pieces.Piece;
 import com.example.jarchess.match.result.ChessMatchResult;
-import com.example.jarchess.match.styles.chessboard.ChessboardStyle;
-import com.example.jarchess.match.styles.chesspiece.ChesspieceStyle;
+import com.example.jarchess.jaraccount.styles.chessboard.ChessboardStyle;
+import com.example.jarchess.jaraccount.styles.chesspiece.ChesspieceStyle;
 import com.example.jarchess.match.turn.Turn;
 
 import java.util.Collection;
