+++ app/src/main/java/com/example/jarchess/online/OnlineMatchInfoBundle.java
@@ -4,8 +4,8 @@ import android.util.Log;
 
 import com.example.jarchess.RemoteOpponentInfoBundle;
 import com.example.jarchess.jaraccount.JarAccount;
-import com.example.jarchess.match.styles.avatar.AvatarStyle;
-import com.example.jarchess.match.styles.avatar.PlayerAvatarStyles;
+import com.example.jarchess.jaraccount.styles.avatar.AvatarStyle;
+import com.example.jarchess.jaraccount.styles.avatar.PlayerAvatarStyles;
 import com.example.jarchess.online.move.DatapackageQueue;
 import com.example.jarchess.online.networking.GameIO;
